move benchmark flow cookbook components from views folder
