+++ app/benchmarking/components/__tests__/cookbooksSelection.test.tsx
@@ -1,8 +1,8 @@
 import { render, screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
+import { CookbooksSelection } from '@/app/benchmarking/components/cookbooksSelection';
 import { CookbooksProvider } from '@/app/benchmarking/contexts/cookbooksContext';
 import { useGetCookbooksQuery } from '@/app/services/cookbook-api-service';
-import { CookbooksSelection } from '@/app/views/cookbook-management/cookbooksSelection';
 import {
   addBenchmarkCookbooks,
   removeBenchmarkCookbooks,
