--- app/views/cookbook-management/hooks/useCookbookList.ts
@@ -1,10 +1,0 @@
-import { useGetCookbooksQuery } from '@/app/services/cookbook-api-service';
-
-export default function useCookbookList() {
-  const { data, error, isLoading, refetch } = useGetCookbooksQuery(undefined);
-  let cookbooks: Cookbook[] = [];
-  if (data !== undefined) {
-    cookbooks = data;
-  }
-  return { cookbooks, error, isLoading, refetch };
-}
